fix: Repository JPA 상속(Integer에서 Long으로 변경)

+++ backend/src/main/java/c104/sinbi/domain/receiver/repository/ReceiverRepository.java
@@ -3,5 +3,5 @@ package c104.sinbi.domain.receiver.repository;
 import c104.sinbi.domain.receiver.Receiver;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReceiverRepository extends JpaRepository<Receiver, Integer> {
+public interface ReceiverRepository extends JpaRepository<Receiver, Long> {
 }
