+++ backend/src/main/java/c104/sinbi/domain/virtualaccount/repository/VirtualAccountRepository.java
@@ -3,5 +3,5 @@ package c104.sinbi.domain.virtualaccount.repository;
 import c104.sinbi.domain.virtualaccount.VirtualAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VirtualAccountRepository extends JpaRepository<VirtualAccount, Integer> {
+public interface VirtualAccountRepository extends JpaRepository<VirtualAccount, Long> {
 }
